backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python
pillow
torch
torchvision
matplotlib
scikit-learn
pandas
seaborn
wget
